pyproject.toml
[project]
name = "scheduling-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "googlemaps>=4.10.0",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.35.0",
]
